bunch of updates * remove regions    blocks and blocks within functions is simpler, which is better for this project.
* lifetime shenanigans for RewriteRuleSet with a Cursor

* unique ids for blocks

* better formatting for Operation & Block
