+++ dialects/arith/src/lib.rs
@@ -4,7 +4,7 @@ pub fn negate<V: Into<Value>>(val: V) -> Operation {
     Operation {
         name: "arith.neg".to_owned(),
         operands: vec![val.into()],
-        regions: Vec::new(),
+        blocks: Vec::new(),
         result: Some(Var::new()),
     }
 }
