+++ dialects/func/src/lib.rs
@@ -1,10 +1,10 @@
-use lorax::{Operation, Region, Value};
+use lorax::{Block, Operation, Value};
 
-pub fn func(region: Region) -> Operation {
+pub fn func(block: Block) -> Operation {
     Operation {
         name: "fn".to_owned(),
         operands: Vec::new(),
-        regions: vec![region],
+        blocks: vec![block],
         result: None,
     }
 }
@@ -13,7 +13,7 @@ pub fn ret<V: Into<Value>>(val: V) -> Operation {
     Operation {
         name: "ret".to_owned(),
         operands: vec![val.into()],
-        regions: Vec::new(),
+        blocks: Vec::new(),
         result: None,
     }
 }
