+++ dialects/x86/src/ops.rs
@@ -4,7 +4,7 @@ pub fn mov(src: Value, dst: Var) -> Operation {
     Operation {
         name: "x86.mov".to_owned(),
         operands: vec![src],
-        regions: Vec::new(),
+        blocks: Vec::new(),
         result: Some(dst),
     }
 }
@@ -13,7 +13,7 @@ pub fn neg(src: Var) -> Operation {
     Operation {
         name: "x86.neg".to_owned(),
         operands: vec![src.into()],
-        regions: Vec::new(),
+        blocks: Vec::new(),
         result: None,
     }
 }
