+++ lorax/src/lib.rs
@@ -3,4 +3,4 @@ mod rewrite;
 mod shape;
 
 pub use rewrite::{Cursor, RewriteRule, RewriteRuleSet};
-pub use shape::{Block, Constant, OpResult, Operation, Region, Value, Var};
+pub use shape::{Block, Constant, OpResult, Operation, Value, Var};
