+++ src/parser/lower.rs
@@ -1,11 +1,11 @@
 // Lower AST to IR
 
-use lorax::{Block, Constant, Region, Value};
+use lorax::{Block, Constant, Value};
 
 use super::ast;
 
 use arith;
-use func;
+use func::{func, ret};
 
 fn lower_expr(block: &mut Block, expr: &ast::Expr) -> Value {
     let op = match expr {
@@ -22,20 +22,21 @@ fn lower_expr(block: &mut Block, expr: &ast::Expr) -> Value {
 
 pub fn lower_stmt(block: &mut Block, stmt: &ast::Stmt) {
     let op = match stmt {
-        ast::Stmt::Return(expr) => func::ret(lower_expr(block, expr)),
+        ast::Stmt::Return(expr) => ret(lower_expr(block, expr)),
     };
 
     block.push(op);
 }
 
-pub fn lower_program(program: &ast::Program) -> Region {
-    let mut region = Region::new();
+pub fn lower_program(program: &ast::Program) -> Block {
+    let mut region = Block::new();
 
     match &program.body {
         ast::Decl::Function(_, stmt) => {
             let mut block = Block::new();
+
             lower_stmt(&mut block, &stmt);
-            region.push(block);
+            region.push(func(block));
         }
     };
